refactor: rename topic resolver, mv to pkg/messaging
