+++ pkg/messaging/routes.go
@@ -1,7 +1,7 @@
 // Copyright (c) Abstract Machines
 // SPDX-License-Identifier: Apache-2.0
 
-package routes
+package messaging
 
 import (
 	"context"
@@ -9,8 +9,8 @@ import (
 	grpcChannelsV1 "github.com/absmach/supermq/api/grpc/channels/v1"
 	grpcCommonV1 "github.com/absmach/supermq/api/grpc/common/v1"
 	grpcDomainsV1 "github.com/absmach/supermq/api/grpc/domains/v1"
-	api "github.com/absmach/supermq/api/http"
 	"github.com/absmach/supermq/pkg/errors"
+	"github.com/gofrs/uuid/v5"
 )
 
 var (
@@ -19,7 +19,7 @@ var (
 	ErrFailedResolveChannel = errors.New("failed to resolve channel route")
 )
 
-type Resolver interface {
+type TopicResolver interface {
 	Resolve(ctx context.Context, domain, channel string) (domainID string, channelID string, err error)
 }
 
@@ -28,7 +28,7 @@ type resolver struct {
 	domains  grpcDomainsV1.DomainsServiceClient
 }
 
-func NewResolver(channelsClient grpcChannelsV1.ChannelsServiceClient, domainsClient grpcDomainsV1.DomainsServiceClient) Resolver {
+func NewTopicResolver(channelsClient grpcChannelsV1.ChannelsServiceClient, domainsClient grpcDomainsV1.DomainsServiceClient) TopicResolver {
 	return &resolver{
 		channels: channelsClient,
 		domains:  domainsClient,
@@ -53,7 +53,7 @@ func (r *resolver) Resolve(ctx context.Context, domain, channel string) (string,
 }
 
 func (r *resolver) resolveDomain(ctx context.Context, domain string) (string, error) {
-	if api.ValidateUUID(domain) == nil {
+	if validateUUID(domain) == nil {
 		return domain, nil
 	}
 	d, err := r.domains.RetrieveByRoute(ctx, &grpcCommonV1.RetrieveByRouteReq{
@@ -67,7 +67,7 @@ func (r *resolver) resolveDomain(ctx context.Context, domain string) (string, er
 }
 
 func (r *resolver) resolveChannel(ctx context.Context, channel, domainID string) (string, error) {
-	if api.ValidateUUID(channel) == nil {
+	if validateUUID(channel) == nil {
 		return channel, nil
 	}
 	c, err := r.channels.RetrieveByRoute(ctx, &grpcCommonV1.RetrieveByRouteReq{
@@ -79,4 +79,13 @@ func (r *resolver) resolveChannel(ctx context.Context, channel, domainID string)
 	}
 
 	return c.Entity.Id, nil
+}
+
+func validateUUID(extID string) (err error) {
+	id, err := uuid.FromString(extID)
+	if id.String() != extID || err != nil {
+		return err
+	}
+
+	return nil
 }
