--- pkg/routes/doc.go
@@ -1,6 +1,0 @@
-// Copyright (c) Abstract Machines
-// SPDX-License-Identifier: Apache-2.0
-
-// Package route contains SuperMQ definition for resolving domain and channel route
-// parts in messaging topics.
-package routes
