+++ pkg/sdk/message_test.go
@@ -28,8 +28,8 @@ import (
 	authnmocks "github.com/absmach/supermq/pkg/authn/mocks"
 	"github.com/absmach/supermq/pkg/errors"
 	svcerr "github.com/absmach/supermq/pkg/errors/service"
+	"github.com/absmach/supermq/pkg/messaging"
 	pubsub "github.com/absmach/supermq/pkg/messaging/mocks"
-	"github.com/absmach/supermq/pkg/routes"
 	sdk "github.com/absmach/supermq/pkg/sdk"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
@@ -47,7 +47,7 @@ func setupMessages() (*httptest.Server, *pubsub.PubSub) {
 	domainsGRPCClient = new(dmocks.DomainsServiceClient)
 	pub := new(pubsub.PubSub)
 	authn := new(authnmocks.Authentication)
-	resolver := routes.NewResolver(channelsGRPCClient, domainsGRPCClient)
+	resolver := messaging.NewTopicResolver(channelsGRPCClient, domainsGRPCClient)
 	handler := adapter.NewHandler(pub, authn, clientsGRPCClient, channelsGRPCClient, resolver, smqlog.NewMock())
 
 	mux := api.MakeHandler(smqlog.NewMock(), "")
